ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
ultralytics==8.0.206
sqlalchemy==2.0.23
opencv-python==4.8.1.78
slowapi==0.1.9
websockets==12.0
